--- a/nixos/lib/testing.nix
+++ b/nixos/lib/testing.nix
@@ -70,8 +70,6 @@ in rec {
             mv $i $out/coverage-data/$(dirname $(dirname $i))
           done
         '';
-
-        meta.needsVMSupport = true;
     };