--- a/nixos/tests/all-tests.nix
+++ b/nixos/tests/all-tests.nix
@@ -2,7 +2,7 @@
 # The return value of this function will be an attrset with arbitrary depth and
 # the `anything` returned by callTest at its test leafs.
 # The tests not supported by `system` will be replaced with `{}`, so that
-# `meta.tests` can contain links to those without breaking on architectures
+# `passthru.tests` can contain links to those without breaking on architectures
 # where said tests are unsupported.
 # Example callTest that just extracts the derivation from the test:
 #   callTest = t: t.test;