--- a/pkgs/servers/mail/opensmtpd/default.nix
+++ b/pkgs/servers/mail/opensmtpd/default.nix
@@ -61,11 +61,8 @@ stdenv.mkDerivation rec {
     license = licenses.isc;
     platforms = platforms.linux;
     maintainers = with maintainers; [ rickynils obadz ekleog ];
-    # https://github.com/NixOS/nixpkgs/issues/50230 must be resolved before
-    # this can be enabled.
-    #
-    # tests = {
-    #   basic-functionality-and-dovecot-interaction = nixosTests.opensmtpd;
-    # };
+  };
+  passthru.tests = {
+    basic-functionality-and-dovecot-interaction = nixosTests.opensmtpd;
   };
 }