… doc/meta.xml
… nixos/tests/run-in-machine.nix
… nixos/tests/sddm.nix
… nixos/tests/virtualbox.nix