This will be easier reviewed commit-by-commit, I tried to add in useful commit messages.

I think this should handle the concerns raised at https://github.com/NixOS/nixpkgs/pull/44439 and https://github.com/NixOS/nixpkgs/issues/50230.

For those who want to read only the commit messages, here they are:

#### tests: refactor to carry the package set as an argument

This way, the package set will be possible to pass without re-importing
all the time

####  tests: split into a separate `all-tests.nix` file

This will make the list much easier to re-use, eg. for `nixosTests`

The drawback is that this approaches makes the
```
nix-build release.nix -A tests.opensmtpd.x86_64-linux
```
command about twice as slow (3s to 6s): it now has to evaluate `nixpkgs`
once for each architecture, instead of just having the hardcoded list of
tests that allowed to say “ok just evaluate for x86_64-linux”.

On the other hand, complete evaluation of `release.nix` should be much
faster because we no longer import `nixpkgs` for each test: testing with
the following command went from 30s to 18s, and that's just for a few
tests.
```
time nix-instantiate --eval --strict nixos/release.nix -A tests.nat
```
I initially wanted to test on the whole `release.nix`, but there are too
many broken tests and it takes too long to eval them all, especially
compared to the fact that the current implementation breaks some setup.

Given developers can just `nix-build nixos/tests/my-test.nix`, it sounds
like an overall win.

#### tests: disable some broken tests and/or restrict to x86_64 

####  meta.tests: drop `meta.needsVMSupport`

Its job is already handled by `requiredSystemFeatures`

####  meta.tests: rename into passthru.tests

Nix currently rejects derivations in `meta` values. This works around
that limitation by using `passthru` instead.

Closes #50230

#### nixosTests: use the newly-extracted all-tests.nix