@Ekleog 2018-11-11
@timokau The `meta -> passthru` is not that urgent: currently the `meta.tests` lines are commented-out. Also, this is fixing the issues @Ericson2314 mentioned, and I don't think we'd have merged the initial PR if we had known them?

Then I must say I'm currently basically thinking “well, this already 3 months, can wait 1 more”, though I'm aware this will merge-conflict as soon as someone adds a test.

Anyway, I can split the “drop meta.needsVMSupport” and “rename to passthru.tests” commits in a separate PR if you'd rather, I was just thinking it'd be better to have nothing than a broken setup, in part because it encourages people to actually review it instead of thinking something along the lines of “well, it works good enough, there are more pressing PRs” :)

BTW let's check whether it actually works with ofborg.

@GrahamcOfBorg build opensmtpd.tests