@GrahamcOfBorg 2018-11-11
<!--REQUEST_ID=a3a114b2-7d76-487c-8a28-0d7ebf8c2b8c-->
Success on x86_64-linux [(full log)](https://logs.nix.ci/?key=nixos/nixpkgs.50233&attempt_id=9d6f1d3c-552c-4cbe-97c3-0ea9545cee19)

Attempted: opensmtpd.tests

<details><summary>Partial log (click to expand)</summary><p>

```
client: running command: sync
client: exit status 0
test script finished in 41.94s
cleaning up
killing smtp1 (pid 597)
killing smtp2 (pid 609)
killing client (pid 621)
vde_switch: EOF on stdin, cleaning up and exiting
vde_switch: Could not remove ctl dir '/build/vde1.ctl': Directory not empty
/nix/store/k46cncv2yvkk1r1lifwkjs9pyxv7qj3h-vm-test-run-opensmtpd
```
</p></details>