@GrahamcOfBorg 2018-11-11
<!--REQUEST_ID=a3a114b2-7d76-487c-8a28-0d7ebf8c2b8c-->
Success on aarch64-linux [(full log)](https://logs.nix.ci/?key=nixos/nixpkgs.50233&attempt_id=b3715b70-6744-43a8-a912-f065d75362ee)

Attempted: opensmtpd.tests

<details><summary>Partial log (click to expand)</summary><p>

```
smtp1: running command: sync
smtp1: exit status 0
test script finished in 111.09s
cleaning up
killing client (pid 631)
killing smtp2 (pid 643)
killing smtp1 (pid 657)
vde_switch: EOF on stdin, cleaning up and exiting
vde_switch: Could not remove ctl dir '/build/vde1.ctl': Directory not empty
/nix/store/drz0xhw9pw7hyb7qqqxz2appcpdnlyqk-vm-test-run-opensmtpd
```
</p></details>