@timokau 2018-11-11
> Then I must say I'm currently basically thinking “well, this already 3 months, can wait 1 more”, though I'm aware this will merge-conflict as soon as someone adds a test.

Sure, whatever way you think works best. I know it can be very frustrating when you spend time improving something and then it goes stale forever. Bonus stress points if there are issues after merge. That happens and is hard to avoid without better tooling. I appreciate your efforts to improve the testing infrastructure :)