@samueldr 2018-11-14
```
env -i nix-build -I . --option restrict-eval true ./nixos/release.nix -A tests.openssh
error: access to path '/nix/store/x29rrys4gmskxpzngwibnmriigycnb79-nixpkgs' is forbidden in restricted mode
(use '--show-trace' to show detailed location information)
```

This currently stops ofborg from testing tests, and I would hazard a guess it will also fail on Hydra.