@Ekleog 2018-11-14
A first attempt at figuring out this is at https://github.com/NixOS/nixpkgs/pull/50186#issuecomment-438272329