@Ekleog 2018-11-14
@samueldr https://github.com/NixOS/nixpkgs/pull/50342 should solve this issue.