@Mic92 2018-11-15
nixops breakage documented here: #50419